+++ src/TopicModel/HLDAModel.hpp
@@ -650,9 +650,9 @@ namespace tomoto
 			return this->globalState.nt->nodes[tid].level;
 		}
 
-		std::vector<size_t> getChildTopicId(Tid tid) const override
+		std::vector<uint32_t> getChildTopicId(Tid tid) const override
 		{
-			std::vector<size_t> ret;
+			std::vector<uint32_t> ret;
 			if (!isLiveTopic(tid)) return ret;
 			for (auto* node = this->globalState.nt->nodes[tid].getChild(); node; node = node->getSibling())
 			{
